moved components into separate files
