+++ code/client/src/index.js
@@ -1,9 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { ResponsivePlayer } from "./player";
+// import { ResponsivePlayer } from "./player";
 import { Header } from "./header";
 import "./css/output.css";
-import { MainBody } from "./body";
+import { Body } from "./body";
 import { Footer } from "./footer";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
@@ -11,7 +11,7 @@ root.render(
   <React.StrictMode>
     <div className="font-sans ">
       <Header />
-      <MainBody />
+      <Body />
       <Footer />
     </div>
   </React.StrictMode>
