+++ code/client/src/left-nav.js
@@ -1,0 +1,19 @@
+export const LeftNav = () => {
+  return (
+    <div>
+      <h3>Table of Contents</h3>
+      <ul>
+        <li>Week 1</li>
+        <li>Week 2</li>
+        <li>Week 3</li>
+        <li>Week 4</li>
+        <li>Week 5</li>
+        <li>Week 6</li>
+        <li>Week 7</li>
+        <li>Week 8</li>
+        <li>Week 9</li>
+        <li>Week 10</li>
+      </ul>
+    </div>
+  );
+};
