+++ code/client/src/player.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import ReactPlayer from "react-player";
+// import ReactPlayer from "react-player";
 
 export class ResponsivePlayer extends Component {
   render() {
