+++ code/client/src/text-box.js
@@ -1,0 +1,78 @@
+export const TextBox = () => {
+  return (
+    <div className="container mx-auto mt-3 p-3 max-h-80 overflow-x-auto rounded-md border-2 border-gray-300 text-sm text-gray-500 ">
+      <p className="m-2">
+        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+        tempor incididunt ut labore et dolore magna aliqua. In ornare quam
+        viverra orci sagittis. Eu consequat ac felis donec et. Leo duis ut diam
+        quam nulla porttitor massa id. Pharetra magna ac placerat vestibulum
+        lectus mauris. In tellus integer feugiat scelerisque varius morbi enim
+        nunc. Nulla aliquet enim tortor at auctor urna nunc id. Massa eget
+        egestas purus viverra accumsan in nisl nisi scelerisque. Velit laoreet
+        id donec ultrices tincidunt arcu non sodales. Sapien et ligula
+        ullamcorper malesuada proin libero nunc consequat interdum. Sem viverra
+        aliquet eget sit amet tellus cras. Nibh praesent tristique magna sit
+        amet purus. Pharetra pharetra massa massa ultricies mi. Arcu cursus
+        vitae congue mauris rhoncus. Sem integer vitae justo eget magna
+        fermentum.
+      </p>
+      <p className="m-2">
+        Integer quis auctor elit sed vulputate. Dui sapien eget mi proin sed
+        libero enim. Odio pellentesque diam volutpat commodo sed egestas egestas
+        fringilla. Convallis posuere morbi leo urna molestie at elementum eu
+        facilisis. Scelerisque mauris pellentesque pulvinar pellentesque
+        habitant morbi tristique senectus et. Facilisis sed odio morbi quis
+        commodo odio. Ut tortor pretium viverra suspendisse potenti nullam ac
+        tortor. Vitae aliquet nec ullamcorper sit amet. Lectus vestibulum mattis
+        ullamcorper velit sed ullamcorper morbi tincidunt. Vitae et leo duis ut
+        diam quam nulla porttitor massa. Sed arcu non odio euismod lacinia.
+        Malesuada nunc vel risus commodo viverra maecenas accumsan lacus vel.
+        Morbi enim nunc faucibus a pellentesque sit amet porttitor eget. Mauris
+        commodo quis imperdiet massa tincidunt nunc pulvinar. In fermentum
+        posuere urna nec tincidunt praesent semper feugiat nibh.
+      </p>
+      <p className="m-2">
+        Diam phasellus vestibulum lorem sed. Laoreet suspendisse interdum
+        consectetur libero id faucibus. Non arcu risus quis varius quam quisque
+        id diam vel. Faucibus scelerisque eleifend donec pretium vulputate
+        sapien nec sagittis. Lacus vel facilisis volutpat est velit egestas dui
+        id. Non sodales neque sodales ut etiam sit amet nisl. Neque gravida in
+        fermentum et. Ut lectus arcu bibendum at varius. Sed odio morbi quis
+        commodo odio aenean sed adipiscing diam. Arcu felis bibendum ut
+        tristique et egestas quis ipsum suspendisse. Eu feugiat pretium nibh
+        ipsum consequat nisl vel pretium. Mattis aliquam faucibus purus in massa
+        tempor nec feugiat. Blandit libero volutpat sed cras ornare arcu dui
+        vivamus. Vitae tempus quam pellentesque nec nam aliquam sem et. Porta
+        non pulvinar neque laoreet suspendisse interdum. Massa id neque aliquam
+        vestibulum morbi. Netus et malesuada fames ac. Urna neque viverra justo
+        nec ultrices dui. Faucibus interdum posuere lorem ipsum.
+      </p>
+      <p className="m-2">
+        Elit at imperdiet dui accumsan sit. Diam donec adipiscing tristique
+        risus nec feugiat in fermentum posuere. Porttitor rhoncus dolor purus
+        non enim praesent elementum. Aliquet eget sit amet tellus cras
+        adipiscing enim. Purus ut faucibus pulvinar elementum integer enim
+        neque. Fermentum et sollicitudin ac orci phasellus egestas tellus.
+        Interdum varius sit amet mattis vulputate. Duis at tellus at urna
+        condimentum. Semper feugiat nibh sed pulvinar proin gravida. Eu
+        consequat ac felis donec et.
+      </p>
+      <p className="m-2">
+        Feugiat in fermentum posuere urna nec tincidunt praesent semper feugiat.
+        Risus feugiat in ante metus dictum at tempor commodo ullamcorper. Enim
+        nunc faucibus a pellentesque sit amet porttitor. Viverra orci sagittis
+        eu volutpat odio facilisis mauris sit. Ac felis donec et odio
+        pellentesque diam volutpat commodo. Facilisis volutpat est velit
+        egestas. Aliquam sem et tortor consequat id porta nibh venenatis. Vitae
+        elementum curabitur vitae nunc. A erat nam at lectus. Lectus quam id leo
+        in. Vulputate mi sit amet mauris commodo quis. Nisl vel pretium lectus
+        quam id leo. Mi bibendum neque egestas congue quisque egestas diam. Amet
+        dictum sit amet justo donec. Vel elit scelerisque mauris pellentesque
+        pulvinar pellentesque habitant morbi. Aliquam ut porttitor leo a diam
+        sollicitudin tempor id eu. Eget nulla facilisi etiam dignissim diam quis
+        enim lobortis scelerisque. Aenean euismod elementum nisi quis eleifend
+        quam adipiscing vitae proin.
+      </p>
+    </div>
+  );
+};
