+++ code/client/src/video-player.js
@@ -1,0 +1,9 @@
+import { ResponsivePlayer } from "./player";
+
+export const VideoPlayer = () => {
+  return (
+    <div className=" mb-3">
+      <ResponsivePlayer />
+    </div>
+  );
+};
